Switch databse engine to sqlite3
Since this application is very read heavy and designed to run on a single machine I’m trying to figure out if I can make it work with SQLite. This would also allow me to more easily offer the dataset to third parties down the road.

The recent rug::b talk and efforts by @fractaledmind pushed me over the edge to give this a try.

To get this going I had to squash the migrations and remove PostgreSQL specific indexes.